Fix the incorrect migration place, add some comment
